Updated test data to dbSNP150 format (using chicken DB)

+++ dbsnp-importer/src/main/java/uk/ac/ebi/eva/dbsnpimporter/io/readers/SubSnpCoreFieldsRowMapper.java
@@ -62,8 +62,8 @@ public class SubSnpCoreFieldsRowMapper implements RowMapper<SubSnpCoreFields> {
     @Override
     public SubSnpCoreFields mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         return new SubSnpCoreFields(
-                resultSet.getInt(SUBSNP_ID_COLUMN),
-                resultSet.getObject(REFSNP_ID_COLUMN, Integer.class),
+                resultSet.getLong(SUBSNP_ID_COLUMN),
+                resultSet.getObject(REFSNP_ID_COLUMN, Long.class),
                 resultSet.getInt(SNP_ORIENTATION_COLUMN),
                 resultSet.getString(CONTIG_NAME_COLUMN),
                 resultSet.getInt(CONTIG_START_COLUMN),
